+++ fractals/src/main/java/me/catmousedog/fractals/fractals/types/normalized/NormalizedInverseMandelbrot.java
@@ -4,21 +4,26 @@ import me.catmousedog.fractals.fractals.Fractal;
 import me.catmousedog.fractals.fractals.filters.Filter;
 import me.catmousedog.fractals.fractals.filters.LogPeriodicFilter;
 import me.catmousedog.fractals.main.Settings;
-import me.catmousedog.fractals.ui.Savable;
 
-public class NormalizedMandelbrot extends Fractal implements Savable {
+/**
+ * Number = Integer
+ */
+public final class NormalizedInverseMandelbrot extends Fractal {
 
-	public NormalizedMandelbrot(Settings settings) {
+	public NormalizedInverseMandelbrot(Settings settings) {
 		super(settings);
 	}
 
-	private NormalizedMandelbrot(Settings settings, Fractal fractal) {
+	private NormalizedInverseMandelbrot(Settings settings, Fractal fractal) {
 		super(settings, fractal);
 	}
 
 	@Override
 	public Number get(double cx, double cy) {
-		double x = cx, y = cy;
+		double kx = cx / (cx * cx + cy * cy);
+		double ky = -cy / (cx * cx + cy * cy);
+
+		double x = kx, y = ky;
 		double tx;
 		double t1, t2;
 
@@ -30,13 +35,13 @@ public class NormalizedMandelbrot extends Fractal implements Savable {
 			t1 = x * x;
 			t2 = y * y;
 
-			x = t1 - t2 + cx;
-			y = 2 * tx * y + cy;
+			x = t1 - t2 + kx;
+			y = 2 * tx * y + ky;
 
-			p = Math.log(t1 + t2)/2;
+			p = Math.log(t1 + t2) / 2;
 
 			if (p > bailout) {
-				return i + 1 - Math.log(p) / Math.log(2);
+				return i + 3 - Math.log(p) / Math.log(2);
 			}
 		}
 		return 0;
@@ -44,12 +49,12 @@ public class NormalizedMandelbrot extends Fractal implements Savable {
 
 	@Override
 	public String informalName() {
-		return "Normalized Mandelbrot";
+		return "Normalized I-Mandelbrot";
 	}
 
 	@Override
 	public String fileName() {
-		return "NormalizedMandelbrot";
+		return "NormalizedInverseMandelbrot";
 	}
 
 	@Override
@@ -59,13 +64,13 @@ public class NormalizedMandelbrot extends Fractal implements Savable {
 	}
 
 	@Override
-	public Fractal clone() {
-		return new NormalizedMandelbrot(settings, this);
-	}
-
-	@Override
 	protected void initFilters() {
 		filters = new Filter[] { new LogPeriodicFilter(this) };
 		filter = filters[0];
+	}
+
+	@Override
+	public Fractal clone() {
+		return new NormalizedInverseMandelbrot(settings, this);
 	}
 }